requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via httpx
async-timeout==4.0.3
    # via langchain
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.0
    # via streamlit
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via
    #   -r requirements.in
    #   pydantic
exceptiongroup==1.3.0
    # via anyio
filetype==1.2.0
    # via langchain-google-genai
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
google-ai-generativelanguage==0.6.18
    # via langchain-google-genai
google-api-core[grpc]==2.24.2
    # via google-ai-generativelanguage
google-auth==2.40.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.2.2
    # via sqlalchemy
grpcio==1.72.0rc1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.72.0rc1
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via langsmith
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
langchain==0.3.25
    # via -r requirements.in
langchain-core==0.3.60
    # via
    #   langchain
    #   langchain-google-genai
    #   langchain-text-splitters
langchain-google-genai==2.1.4
    # via -r requirements.in
langchain-text-splitters==0.3.8
    # via langchain
langsmith==0.3.42
    # via
    #   langchain
    #   langchain-core
markupsafe==3.0.2
    # via jinja2
narwhals==1.40.0
    # via altair
numpy==2.2.6
    # via
    #   opencv-python
    #   pandas
    #   pydeck
    #   streamlit
opencv-python==4.11.0.86
    # via -r requirements.in
orjson==3.10.18
    # via langsmith
packaging==24.2
    # via
    #   altair
    #   langchain-core
    #   langsmith
    #   pytesseract
    #   streamlit
pandas==2.2.3
    # via streamlit
pillow==11.2.1
    # via
    #   -r requirements.in
    #   pytesseract
    #   streamlit
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==6.31.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   streamlit
pyarrow==20.0.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic[email]==2.11.4
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-core
    #   langchain-google-genai
    #   langsmith
pydantic-core==2.33.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pytesseract==0.3.13
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-core
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   google-api-core
    #   langchain
    #   langsmith
    #   requests-toolbelt
    #   streamlit
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.25.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.41
    # via langchain
streamlit==1.45.1
    # via -r requirements.in
tenacity==9.1.2
    # via
    #   langchain-core
    #   streamlit
toml==0.10.2
    # via streamlit
tornado==6.5
    # via streamlit
typing-extensions==4.13.2
    # via
    #   altair
    #   anyio
    #   exceptiongroup
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
watchdog==6.0.0
    # via streamlit
zstandard==0.23.0
    # via langsmith
